feat: add user information to '/me' page
This adds basic user information, like account name, admin role indication, and number of upvotes/downvotes to the '/me' page, and links to 'My Ideas' and 'edit profile' pages (they're doubled in the drop down menu, but hopefully it will make navigation easier).
If user is not logged in, the page will show a notification they can't see the page unless they're logged in.